separate filtering functionality into a separate mod. Simplify part of the covcalc
