+++ src/bamcompare.rs
@@ -8,7 +8,8 @@ use std::fs::File;
 use itertools::Itertools;
 use bigtools::{Value};
 use crate::filehandler::{bam_ispaired, write_covfile, is_bed_or_gtf, read_bedfile};
-use crate::covcalc::{bam_pileup, parse_regions, Alignmentfilters, TempZip, region_divider, Region};
+use crate::covcalc::{bam_pileup, parse_regions, TempZip, region_divider, Region};
+use crate::filtering::Alignmentfilters;
 use crate::normalization::scale_factor_bamcompare;
 use crate::calc::{median, calc_ratio};
 use tempfile::{TempPath};
