+++ src/multibamsummary.rs
@@ -9,7 +9,8 @@ use std::fs::File;
 use ndarray::Array2;
 use ndarray_npy::NpzWriter;
 use std::collections::HashMap;
-use crate::covcalc::{bam_pileup, parse_regions, Alignmentfilters, TempZip, region_divider};
+use crate::covcalc::{bam_pileup, parse_regions, TempZip, region_divider};
+use crate::filtering::Alignmentfilters;
 use crate::filehandler::{bam_ispaired, read_bedfile, read_gtffile, chrombounds_from_bam, is_bed_or_gtf};
 use crate::calc::deseq_scalefactors;
 use crate::covcalc::{Region, Gtfparse};
